+++ tests/t30010/libraries/common.h
@@ -1,0 +1,15 @@
+#pragma once
+
+#include "../libraries/lib1/lib1.h"
+#include "../libraries/lib2/lib2.h"
+#include "../libraries/lib3/lib3.h"
+#include "../libraries/lib4/lib4.h"
+
+namespace clanguml {
+namespace t30010 {
+
+enum class Const { TIMEOUT = 1 };
+
+int init();
+}
+}